Introduce and use small[u]int type. Changes in size:
-     79              0      28     107      6b libc/inet/rpc/create_xid.o
+     76              0      25     101      65 libc/inet/rpc/create_xid.o
-    126              0       4     130      82 libc/misc/assert/__assert.o
+    123              0       1     124      7c libc/misc/assert/__assert.o
-    648              4      24     676     2a4 libc/misc/internals/__uClibc_main.o
+    645              4      21     670     29e libc/misc/internals/__uClibc_main.o
-    230              0       4     234      ea libc/stdlib/abort.o
+    216              0       1     217      d9 libc/stdlib/abort.o
-    129              0       4     133      85 libc/termios/tcgetsid.o
+    126              0       1     127      7f libc/termios/tcgetsid.o

+++ libc/inet/rpc/create_xid.c
@@ -35,7 +35,7 @@ libc_hidden_proto(gettimeofday)
 __UCLIBC_MUTEX_STATIC(mylock, PTHREAD_MUTEX_INITIALIZER);
 
 
-static int is_initialized;
+static smallint is_initialized;
 static struct drand48_data __rpc_lrand48_data;
 
 u_long _create_xid (void) attribute_hidden;
